% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_price.R
\name{total_price}
\alias{total_price}
\title{Computes the total price of a order after applying a discount}
\usage{
total_price(data, currency = "EUR")
}
\arguments{
\item{data}{input order dataset}

\item{currency}{"EUR" or "USD"}
}
\value{
A numeric value of the total price of the order
}
\description{
The discount depends on the currency (EUR or USD) and the date.
}
\details{
On 27th November 2020 the discount is 50\%.
On the other dates the discount is 20\% for EUR or 25\% for USD.
}
\examples{
total_price(data = order, var = "EUR")

}
